+++ pkg/parser/ansible/ini/hosts/parser_test.go
@@ -1,4 +1,4 @@
-package ansiblehosts
+package hosts
 
 import (
 	"encoding/json"
